scanner_service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
img2pdf==0.5.1
reportlab==4.0.7
python-sane==2.9.1
